# docs/openapi.yaml
openapi: 3.0.3
info:
  title: PCCraft Marketplace API
  version: 1.0.0
  description: REST API for the PCCraft Marketplace platform -- multi-vendor PC components
    e-commerce with recommendations and a data-driven compatibility engine.
paths:
...
  /api/v1/admin/brands/:
    get:
      operationId: admin_brands_retrieve
      description: '``GET /`` paginated list, ``POST /`` create.'
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    post:
      operationId: admin_brands_create
      description: '``GET /`` paginated list, ``POST /`` create.'
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/v1/admin/brands/{slug}/:
    get:
      operationId: admin_brands_retrieve_2
      description: '``GET / PATCH / DELETE /api/v1/admin/brands/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    patch:
      operationId: admin_brands_partial_update
      description: '``GET / PATCH / DELETE /api/v1/admin/brands/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    delete:
      operationId: admin_brands_destroy
      description: '``GET / PATCH / DELETE /api/v1/admin/brands/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/admin/brands/{slug}/restore/:
    patch:
      operationId: admin_brands_restore_partial_update
      description: '``PATCH /api/v1/admin/brands/{slug}/restore/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/v1/admin/categories/:
    get:
      operationId: admin_categories_retrieve
      description: '``GET /`` paginated list, ``POST /`` create.'
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    post:
      operationId: admin_categories_create
      description: '``GET /`` paginated list, ``POST /`` create.'
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/v1/admin/categories/{slug}/:
    get:
      operationId: admin_categories_retrieve_2
      description: '``GET / PATCH / DELETE /api/v1/admin/categories/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    patch:
      operationId: admin_categories_partial_update
      description: '``GET / PATCH / DELETE /api/v1/admin/categories/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    delete:
      operationId: admin_categories_destroy
      description: '``GET / PATCH / DELETE /api/v1/admin/categories/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/v1/admin/categories/{slug}/restore/:
    patch:
      operationId: admin_categories_restore_partial_update
      description: '``PATCH /api/v1/admin/categories/{slug}/restore/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/v1/admin/categories/tree/:
    get:
      operationId: admin_categories_tree_retrieve
      description: '``GET /api/v1/admin/categories/tree/``.'
      tags:
      - admin
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
  /api/v1/brands/:
    get:
      operationId: brands_retrieve
      description: '``GET /`` public; ``POST /`` admin only.'
      tags:
      - brands
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: No response body
    post:
      operationId: brands_create
      description: '``GET /`` public; ``POST /`` admin only.'
      tags:
      - brands
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/v1/brands/{slug}/:
    get:
      operationId: brands_retrieve_2
      description: '``GET / PATCH / DELETE /api/v1/brands/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - brands
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: No response body
    patch:
      operationId: brands_partial_update
      description: '``GET / PATCH / DELETE /api/v1/brands/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - brands
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    delete:
      operationId: brands_destroy
      description: '``GET / PATCH / DELETE /api/v1/brands/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - brands
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/categories/:
    get:
      operationId: categories_retrieve
      description: '``GET /`` public tree (or flat); ``POST /`` admin-only create.'
      tags:
      - categories
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: No response body
    post:
      operationId: categories_create
      description: '``GET /`` public tree (or flat); ``POST /`` admin-only create.'
      tags:
      - categories
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
  /api/v1/categories/{slug}/:
    get:
      operationId: categories_retrieve_2
      description: '``GET / PATCH / DELETE /api/v1/categories/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - categories
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: No response body
    patch:
      operationId: categories_partial_update
      description: '``GET / PATCH / DELETE /api/v1/categories/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - categories
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    delete:
      operationId: categories_destroy
      description: '``GET / PATCH / DELETE /api/v1/categories/{slug}/``.'
      parameters:
      - in: path
        name: slug
        schema:
          type: string
        required: true
      tags:
      - categories
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/v1/search/:
    get:
      operationId: search_retrieve
      description: Full-text + faceted product search (Module 11 §11.1).
      tags:
      - search
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT